refactored similar tests in a new file
